Add breakpoint into .zshrc
this will inform you while the shell and it's modules are initialzing
when it is a safe time where you can interrupt the initialzing.

"safe time": right before the antigen bundles are beeing loaded.

this is useful if you are about to logon to a box which is experiencing
heavy load and would take ages to initialize all the antigen modules